+++ src/api/controllers/guild/access.js
@@ -6,15 +6,15 @@ import { isAccessAllowed } from 'lib/services/guild';
 import { isUserInGuild } from 'lib/services/user';
 
 type Options = {
-  guildName: string,
   type: string,
-  email: string,
+  guildName?: string,
+  email?: string,
 };
 
 export default async function canAccess (models: Models, { guildName, type, email }: Options) {
   const [publicAccessAllowed, userIsInGuild] = await Promise.all([
     isAccessAllowed(models, type, guildName),
-    isUserInGuild(models, email, guildName),
+    isUserInGuild(models, email || '', guildName || ''),
   ]);
 
   return !!(userIsInGuild || publicAccessAllowed);